Josh's Bar graph edit
idk what I did but I fixed the bar graph. If one of you guys can summarize the countries of origin and visualize that, and one of y'all do the winner and loser rank points I think that would be solid. We also need the summary but I can get started on that pretty soon.