Final Visualization Completed
Ok so I've finished all the visualizations of the graphs we want to do. I did the country data, the rankings, the ranking points, and the average number of aces. If you guys can just write in little text boxes next to the visualizations what the purpose of each is, as well as the written portions, we should be set